Add exception for libassistant to import checker.
libassistant needs to be moved to third_party to fix the actual issue
but adding an exception is a quick fix that prevents constantly
seeing the false positive when working on a repo with Android and
Linux.

This fixes issues with "multiple definition" involving a class and
the same class in libassistant.

Change-Id: I5507bb3b969c03b521107ae3bf64742d13886301
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1802297
Reviewed-by: Ted Choc <tedchoc@chromium.org>
Commit-Queue: Henrique Nakashima <hnakashima@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#696237}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: a70956aaab88a148eb884d639debfe6c8fa8f610