Fix checkdeps
checkdeps assumes that no DEPS rule ends in a slash. If a DEPS rule
ends in a slash, it is ignored and not enforced. This CL fixes this
bug by making DEPS rules that end in a slash an error.

There are five DEPS files that have a rule ending in a slash in the
repo, though only two of them (fortunately!) accidentally committed
files violating the DEPS.

Bugs are filed.

Bug: 1084817, 1084826, 1084827
Change-Id: Ifbfdc4c16b89a36b26df51eee1897d0385689181
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2208299
Reviewed-by: Tommy Nyquist <nyquist@chromium.org>
Reviewed-by: Tarun Bansal <tbansal@chromium.org>
Reviewed-by: Dirk Pranke <dpranke@chromium.org>
Reviewed-by: Devlin <rdevlin.cronin@chromium.org>
Commit-Queue: Avi Drissman <avi@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#770669}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 193fbc60395e70ce013c42dcaad00ab5e77bbbaa