cct: calls setClientDataHeaderForNewTab for hidden tabs
This adds a call to setClientDataHeaderForNewTab() for tabs created
when mayLaunchUrl is called. Without this, the right header is not
added if mayLaunchUrl() was called.

This also adds AppHooksImpl to the list of files that has multiple
paths, and changes the presubmit checking to prefer the public repo
in the case of multiple paths matching. This ensures that includes to
AppHooksImpl don't need to add DEPS to clank.

BUG=1207675
TEST=testMayLaunchUrlAddsClientDataHeader

Change-Id: I2ebbcb1b54ffae50203e0843f71eea5eb40b9fc3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2891644
Reviewed-by: Dirk Pranke <dpranke@google.com>
Reviewed-by: Peter Conn <peconn@chromium.org>
Reviewed-by: Ted Choc <tedchoc@chromium.org>
Commit-Queue: Scott Violet <sky@chromium.org>
Cr-Commit-Position: refs/heads/master@{#883000}
NOKEYCHECK=True
GitOrigin-RevId: 1dbd37e18f9046e2772f6949050675b4cef5054b