[libc++] Add a gn variable to enable libc++ hardening
This gn variable enable_safe_libcxx determines whether or not libc++
will be built with assertion checks for safety. It's default value will
be true for now.

The motivation for this gn arg is that the V8 team raised some concerns
about the performance overhead of enabling libc++ assertions [0], so
adding a gn variable creates an easy way to disable them if necessary
(as opposed to having to revert a bunch of CLs).

[0]: https://crbug.com/1335422#c45

Bug: 1335422
Change-Id: I430d2c8fb0447e220c227a2340eb01984010d85a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3910960
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Nico Weber <thakis@chromium.org>
Commit-Queue: Alan Zhao <ayzhao@google.com>
Cr-Commit-Position: refs/heads/main@{#1064497}
NOKEYCHECK=True
GitOrigin-RevId: a9f87b92eee0f3ddc5a9c9a6dffddf9b9b935dee